Move theme protocols to Common
