+++ Shared/Theme/LightTheme.swift
@@ -3,6 +3,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import UIKit
+import Common
 
 public struct LightTheme: Theme {
     public var type: ThemeType = .light
